requirements.txt
langchain
langgraph
groq
streamlit
reportlab
python-docx
python-dotenv
mysql-connector-python